Fixed a bug in the initialisation of yambopy input file class. Added Boron Nitride to the tutorial
There was a problem in the initialisation of the yambo class where
sometimes it would read a already present yambo.in file. Now by default
we remove the yambo.in before initialising the class. Added a example
for Boron Nitride for BSE in serial and splitting the calculation of
the dielectric function at different q points in different jobs. If the
output folder is always the same yambo might mix up the q points and
give wrong results without warning. The file bse_bug.png shows this
problem.