Removed all references to spur
It seemed a good idea to include spur in the tutorial because it allows
to run yambo remotely via ssh. Since this feature wouldn’t work in the
way it is intended and spur is not trivial to install in some cases, I
removed all the references to it on the tutorial.